pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bioimage-mcp"
version = "0.0.0"
description = "Bioimage-MCP: local-first MCP server for bioimage tools"
readme = "README.md"
requires-python = ">=3.13"
license = { file = "LICENSE" }
authors = [{ name = "OpenAI Codex" }]
dependencies = [
  "mcp",
  "pydantic>=2",
  "PyYAML",
  "bioio",
  "bioio-ome-zarr",
  "bioio-ome-tiff",
  "ngff-zarr",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "ruff>=0.6",
]

[project.scripts]
bioimage-mcp = "bioimage_mcp.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

